app/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1
slowapi==0.1.9
openpyxl==3.1.5
python-multipart==0.0.18
statsmodels>=0.14
